pyproject.toml
[project]
name = "rjm-backend-core"
version = "0.1.0"
description = "RJM MIRA backend core API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]>=0.120.0",
    "loguru>=0.7.3",
    "pytest>=8.4.2",
    "sqlmodel>=0.0.16",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.20.0",
    "greenlet>=3.0.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.3.0",
    "python-jose[cryptography]>=3.3.0",
    "openai>=1.0.0",
    "pinecone>=4.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]
